Fix backup overriding at rendering.

+++ newIDE/app/src/Utils/SerializableObjectCancelableEditor.js
@@ -179,22 +179,24 @@ export const useSerializableObjectsCancelableEditor = ({
   const backdropClickBehavior = preferences.values.backdropClickBehavior;
 
   const serializedElements = serializedElementsRef.current;
-  for (const [id, serializableObject] of serializableObjects) {
-    // Serialize the content of the object, to be used in case the user
-    // want to cancel their changes.
-    {
-      const serializedElement = serializedElements.get(id);
-      if (serializedElement) {
-        serializedElement.delete();
-        serializedElements.delete(id);
+  if (serializedElements.size === 0) {
+    for (const [id, serializableObject] of serializableObjects) {
+      // Serialize the content of the object, to be used in case the user
+      // want to cancel their changes.
+      {
+        const serializedElement = serializedElements.get(id);
+        if (serializedElement) {
+          serializedElement.delete();
+          serializedElements.delete(id);
+        }
       }
+      if (resetThenClearPersistentUuid) {
+        serializableObject.resetPersistentUuid();
+      }
+      const serializedElement = new gd.SerializerElement();
+      serializableObject.serializeTo(serializedElement);
+      serializedElements.set(id, serializedElement);
     }
-    if (resetThenClearPersistentUuid) {
-      serializableObject.resetPersistentUuid();
-    }
-    const serializedElement = new gd.SerializerElement();
-    serializableObject.serializeTo(serializedElement);
-    serializedElements.set(id, serializedElement);
   }
 
   React.useEffect(
